IEEE slow: Use %zu to print sizeof values

+++ print-slow.c
@@ -382,7 +382,7 @@ slow_marker_lacp_print(netdissect_options *ndo,
         /* length includes the type and length fields */
         if (tlv_len < sizeof(struct tlv_header_t)) {
             ND_PRINT("\n\t    ERROR: illegal length - should be >= %zu",
-                   sizeof(struct tlv_header_t));
+                     sizeof(struct tlv_header_t));
             return;
         }
 
@@ -403,7 +403,7 @@ slow_marker_lacp_print(netdissect_options *ndo,
             if (tlv_tlen !=
                 sizeof(struct lacp_tlv_actor_partner_info_t)) {
                 ND_PRINT("\n\t    ERROR: illegal length - should be %zu",
-                       sizeof(struct tlv_header_t) + sizeof(struct lacp_tlv_actor_partner_info_t));
+                         sizeof(struct tlv_header_t) + sizeof(struct lacp_tlv_actor_partner_info_t));
                 goto badlength;
             }
 
@@ -426,7 +426,7 @@ slow_marker_lacp_print(netdissect_options *ndo,
             if (tlv_tlen !=
                 sizeof(struct lacp_tlv_collector_info_t)) {
                 ND_PRINT("\n\t    ERROR: illegal length - should be %zu",
-                       sizeof(struct tlv_header_t) + sizeof(struct lacp_tlv_collector_info_t));
+                         sizeof(struct tlv_header_t) + sizeof(struct lacp_tlv_collector_info_t));
                 goto badlength;
             }
 
@@ -441,7 +441,7 @@ slow_marker_lacp_print(netdissect_options *ndo,
             if (tlv_tlen !=
                 sizeof(struct marker_tlv_marker_info_t)) {
                 ND_PRINT("\n\t    ERROR: illegal length - should be %zu",
-                       sizeof(struct tlv_header_t) + sizeof(struct marker_tlv_marker_info_t));
+                         sizeof(struct tlv_header_t) + sizeof(struct marker_tlv_marker_info_t));
                 goto badlength;
             }
 
@@ -506,7 +506,7 @@ slow_oam_print(netdissect_options *ndo,
     } ptr;
 
     union {
-	const struct slow_oam_info_t *slow_oam_info;
+        const struct slow_oam_info_t *slow_oam_info;
         const struct slow_oam_link_event_t *slow_oam_link_event;
         const struct slow_oam_variablerequest_t *slow_oam_variablerequest;
         const struct slow_oam_variableresponse_t *slow_oam_variableresponse;
@@ -551,8 +551,8 @@ slow_oam_print(netdissect_options *ndo,
 
             /* length includes the type and length fields */
             if (length < sizeof(struct slow_oam_tlv_header_t)) {
-                ND_PRINT("\n\t    ERROR: illegal length - should be >= %u",
-                       (u_int)sizeof(struct slow_oam_tlv_header_t));
+                ND_PRINT("\n\t    ERROR: illegal length - should be >= %zu",
+                         sizeof(struct slow_oam_tlv_header_t));
                 return;
             }
 
@@ -649,8 +649,8 @@ slow_oam_print(netdissect_options *ndo,
 
             /* length includes the type and length fields */
             if (length < sizeof(struct slow_oam_tlv_header_t)) {
-                ND_PRINT("\n\t    ERROR: illegal length - should be >= %u",
-                       (u_int)sizeof(struct slow_oam_tlv_header_t));
+                ND_PRINT("\n\t    ERROR: illegal length - should be >= %zu",
+                         sizeof(struct slow_oam_tlv_header_t));
                 return;
             }
 
@@ -669,7 +669,7 @@ slow_oam_print(netdissect_options *ndo,
                 if (GET_U_1(tlv.slow_oam_link_event->event_length) !=
                     sizeof(struct slow_oam_link_event_t)) {
                     ND_PRINT("\n\t    ERROR: illegal length - should be %zu",
-                           sizeof(struct slow_oam_link_event_t));
+                             sizeof(struct slow_oam_link_event_t));
                     hexdump = TRUE;
                     goto badlength_event_notif;
                 }
